[key] move keys related functions and types in keys
